refactor: MenuProduct quantity 포장
