+++ src/test/java/kitchenpos/menu/domain/MenuProductTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
 import kitchenpos.common.vo.Price;
+import kitchenpos.common.vo.Quantity;
 import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -14,8 +15,8 @@ class MenuProductTest {
     @Test
     void construct() {
         Product 타코야끼 = new Product(1L, "타코야끼", new Price(BigDecimal.valueOf(12000)));
-        MenuProduct menuProduct = new MenuProduct(타코야끼, 2);
-        MenuProduct expectedMenuProduct = new MenuProduct(타코야끼, 2);
+        MenuProduct menuProduct = new MenuProduct(타코야끼, new Quantity(2));
+        MenuProduct expectedMenuProduct = new MenuProduct(타코야끼, new Quantity(2));
         assertThat(menuProduct.getProduct()).isEqualTo(expectedMenuProduct.getProduct());
     }
 }
